txring and drain logic for outgoing packets

+++ src/xdpd/drivers/gnu_linux_dpdk_ng/src/io/iface_manager.cc
@@ -1240,7 +1240,7 @@ rofl_result_t iface_manager_discover_physical_ports(void){
 
 			/* store txring-drain-interval parameter for this port */
 			if (not phyports[port_id].is_virtual && iface_manager_port_setting_exists(s_pci_addr, "txring-drain-interval")) {
-				tx_core_tasks[lcore_id].txring_drain_interval[port_id] = iface_manager_get_port_setting_as<unsigned int>(s_pci_addr, "txring-drain-interval");
+				tx_core_tasks[lcore_id].txring_drain_interval[port_id] = iface_manager_get_port_setting_as<uint64_t>(s_pci_addr, "txring-drain-interval");
 			} else {
 				tx_core_tasks[lcore_id].txring_drain_interval[port_id] = PROCESSING_TXRING_DRAIN_INTERVAL_DEFAULT;
 			}
